refactor: Cleanup TravelTimeResource
